+++ x/programs/runtime/program.go
@@ -10,7 +10,6 @@ import (
 	"errors"
 
 	"github.com/bytecodealliance/wasmtime-go/v14"
-	"github.com/near/borsh-go"
 
 	"github.com/ava-labs/hypersdk/codec"
 )
@@ -85,7 +84,7 @@ func (p *ProgramInstance) call(_ context.Context, callInfo *CallInfo) ([]byte, e
 
 	// create the program context
 	programCtx := Context{Program: callInfo.Program, Actor: callInfo.Actor}
-	paramsBytes, err := borsh.Serialize(programCtx)
+	paramsBytes, err := serialize(programCtx)
 	if err != nil {
 		return nil, err
 	}
